Format (#2)
* create src, include and cmakelists
* ignore build folder
* add clang format and tidy and format files
* remove from unnecessary folder
