--- ConsoleRPG/STLINCLUDE.h
@@ -1,13 +1,0 @@
-#pragma once
-
-#include<stdlib.h>
-#include<string>
-#include<ctime>
-#include<vector>
-#include<sstream>
-#include<fstream>
-#include<iostream>
-#include<iomanip>
-
-#include"dArr.h"
-#include"gui_functions.h"
