+++ ConsoleRPG/include/Boss.h
@@ -3,10 +3,7 @@
 class Boss
 {
 private:
-
-
 public:
-	Boss();
-	virtual ~Boss();
+  Boss();
+  virtual ~Boss();
 };
-
