+++ ConsoleRPG/include/Functions.h
@@ -1,5 +1,4 @@
 #ifndef FUNCTIONS_H
 #define FUNCTIONS_H
 
-
 #endif
