+++ ConsoleRPG/src/main.cpp
@@ -1,0 +1,16 @@
+#include "Game.h"
+
+int main()
+{
+  srand(time(NULL));
+
+  Game game;
+  game.initGame();
+
+  while (game.getPlaying())
+  {
+    game.mainMenu();
+  }
+
+  return 0;
+}
